UI: Move getErrors to AuthService
