+++ monkey/monkey_island/cc/ui/src/components/pages/RegisterPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Row, Col, Container, Form, Button} from 'react-bootstrap';
 
-import AuthService from '../../services/AuthService';
+import AuthService, {getErrors} from '../../services/AuthService';
 import monkeyDetective from '../../images/detective-monkey.svg';
 import ParticleBackground from '../ui-components/ParticleBackground';
 import LoadingIcon from '../ui-components/LoadingIcon';
@@ -35,16 +35,6 @@ class RegisterPageComponent extends React.Component {
 
   redirectToHome = () => {
     window.location.href = '/landing-page';
-  };
-
-  getErrors = (errors) => {
-    const errorArray = [];
-
-    for (let i=0; i<errors.length; i++) {
-      const key = 'registration-error-' + i
-      errorArray.push(<li key={key}>{errors[i]}</li>);
-    }
-    return <ul>{errorArray}</ul>;
   };
 
   constructor(props) {
@@ -92,7 +82,7 @@ class RegisterPageComponent extends React.Component {
                       <Col>
                         {
                           this.state.failed ?
-                            <div className='alert alert-danger' role='alert'>{this.getErrors(this.state.errors)}</div>
+                            <div className='alert alert-danger' role='alert'>{getErrors(this.state.errors)}</div>
                             :
                             ''
                         }
